+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 - Give an example of using this operator with a Flux GitRepository and webhooks, in
   `examples/flux-source`.
   [#339](https://github.com/pulumi/pulumi-kubernetes-operator/pull/339)
+- De-escalate a log message about a harmless error from ERROR to DEBUG
+  [#352](https://github.com/pulumi/pulumi-kubernetes-operator/pull/352)
 
 ## 1.10.0 (2022-10-21)